amifuse: add Windows auto-mountpoint support
Extend the auto-mountpoint logic to support Windows by automatically
selecting the first available drive letter when --mountpoint is not
specified.

The search iterates through A: to Z: and uses the first drive letter
that doesn't exist. If no drive letter is available, the user is
prompted to specify one manually via --mountpoint.